+++ src/main/java/de/maxhenkel/corpse/entities/DummySkeleton.java
@@ -3,7 +3,7 @@ package de.maxhenkel.corpse.entities;
 import de.maxhenkel.corpse.CorpseMod;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.EquipmentSlot;
-import net.minecraft.world.entity.monster.Skeleton;
+import net.minecraft.world.entity.monster.skeleton.Skeleton;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.Level;