+++ src/main/java/de/maxhenkel/corpse/gui/ScreenBase.java
@@ -4,11 +4,11 @@ import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.client.gui.screens.Screen;
 import net.minecraft.client.renderer.RenderPipelines;
 import net.minecraft.network.chat.Component;
-import net.minecraft.resources.ResourceLocation;
+import net.minecraft.resources.Identifier;
 
 public class ScreenBase extends Screen {
 
-    protected ResourceLocation texture;
+    protected Identifier texture;
 
     protected int guiLeft;
     protected int guiTop;