+++ src/main/java/de/maxhenkel/corpse/net/MessageShowCorpseInventory.java
@@ -6,7 +6,7 @@ import de.maxhenkel.corpse.gui.Guis;
 import net.minecraft.network.RegistryFriendlyByteBuf;
 import net.minecraft.network.protocol.PacketFlow;
 import net.minecraft.network.protocol.common.custom.CustomPacketPayload;
-import net.minecraft.resources.ResourceLocation;
+import net.minecraft.resources.Identifier;
 import net.minecraft.server.level.ServerPlayer;
 import net.neoforged.neoforge.network.handling.IPayloadContext;
 
@@ -14,7 +14,7 @@ import java.util.UUID;
 
 public class MessageShowCorpseInventory implements Message<MessageShowCorpseInventory> {
 
-    public static final CustomPacketPayload.Type<MessageShowCorpseInventory> TYPE = new CustomPacketPayload.Type<>(ResourceLocation.fromNamespaceAndPath(CorpseMod.MODID, "show_corpse_inventory"));
+    public static final CustomPacketPayload.Type<MessageShowCorpseInventory> TYPE = new CustomPacketPayload.Type<>(Identifier.fromNamespaceAndPath(CorpseMod.MODID, "show_corpse_inventory"));
 
     private UUID playerUUID;
     private UUID deathID;