+++ src/main/java/de/maxhenkel/corpse/net/MessageSpawnDeathParticles.java
@@ -5,14 +5,14 @@ import de.maxhenkel.corpse.CorpseMod;
 import net.minecraft.network.RegistryFriendlyByteBuf;
 import net.minecraft.network.protocol.PacketFlow;
 import net.minecraft.network.protocol.common.custom.CustomPacketPayload;
-import net.minecraft.resources.ResourceLocation;
+import net.minecraft.resources.Identifier;
 import net.neoforged.neoforge.network.handling.IPayloadContext;
 
 import java.util.UUID;
 
 public class MessageSpawnDeathParticles implements Message<MessageSpawnDeathParticles> {
 
-    public static final CustomPacketPayload.Type<MessageSpawnDeathParticles> TYPE = new CustomPacketPayload.Type<>(ResourceLocation.fromNamespaceAndPath(CorpseMod.MODID, "spawn_death_particles"));
+    public static final CustomPacketPayload.Type<MessageSpawnDeathParticles> TYPE = new CustomPacketPayload.Type<>(Identifier.fromNamespaceAndPath(CorpseMod.MODID, "spawn_death_particles"));
 
     private UUID corpseUUID;